# ktoolbox/webui/pawchive_routes.py
from __future__ import annotations

from typing import Annotated, cast

from fastapi import APIRouter, Depends, HTTPException, Request, status
from pydantic import ValidationError

from ktoolbox.action.search import search_creator, search_creator_post
from ktoolbox.api.errors import PawchiveError, PawchiveNotFoundError
from ktoolbox.api.generated import Post, Revision
from ktoolbox.api.utils import create_pawchive_client
from ktoolbox.configuration import RuntimeContext
from ktoolbox.publication_time import PublishedTimeError
from ktoolbox.webui.auth import require_session
from ktoolbox.webui.database import WebUISession
from ktoolbox.webui.media import creator_search_item, post_detail, post_summary, revision_detail, revision_summary
from ktoolbox.webui.models import (
    CreatorSearchItemResponse,
    PawchivePostDetailResponse,
    PawchivePostSummaryResponse,
    PawchiveRevisionDetailResponse,
    PawchiveRevisionSummaryResponse,
    SiteVersionResponse,
)

SessionDependency = Annotated[WebUISession, Depends(require_session)]


def publication_time_http_error(error: PublishedTimeError) -> HTTPException:
    return HTTPException(
        status_code=status.HTTP_502_BAD_GATEWAY,
        detail={
            "code": "response_incompatible",
            "message": "Pawchive returned an invalid publication time",
            "service": error.service,
            "timezone": error.timezone_name,
            "fields": ["published"],
        },
    )


def create_pawchive_router() -> APIRouter:
    router = APIRouter(prefix="/api/v1/pawchive")

    def runtime(request: Request) -> RuntimeContext:
        return cast(RuntimeContext, request.app.state.runtime_context).snapshot()

    def upstream_error(message: str) -> HTTPException:
        return HTTPException(
            status_code=status.HTTP_502_BAD_GATEWAY,
            detail={"code": "upstream_error", "message": message},
        )

    @router.get("/creators", response_model=list[CreatorSearchItemResponse])
    async def creators(
        request: Request,
        _: SessionDependency,
        creator_id: str | None = None,
        name: str | None = None,
        service: str | None = None,
    ) -> list[CreatorSearchItemResponse]:
        context = runtime(request)
        with context.activate():
            async with create_pawchive_client() as client:
                result = await search_creator(id=creator_id, name=name, service=service, client=client)
        if not result:
            raise upstream_error(result.message or "creator search failed")
        return [creator_search_item(creator) for creator in result.data or ()]

    @router.get("/posts", response_model=list[PawchivePostSummaryResponse])
    async def posts(
        request: Request,
        _: SessionDependency,
        creator_id: str | None = None,
        name: str | None = None,
        service: str | None = None,
        query: str | None = None,
        offset: int | None = None,
    ) -> list[PawchivePostSummaryResponse]:
        context = runtime(request)
        with context.activate():
            async with create_pawchive_client() as client:
                result = await search_creator_post(
                    id=creator_id,
                    name=name,
                    service=service,
                    q=query,
                    o=offset,
                    client=client,
                )
        if not result:
            code = (
                status.HTTP_422_UNPROCESSABLE_CONTENT
                if not any((creator_id, name, service))
                else status.HTTP_502_BAD_GATEWAY
            )
            detail_code = "validation_failed" if code == status.HTTP_422_UNPROCESSABLE_CONTENT else "upstream_error"
            raise HTTPException(
                status_code=code,
                detail={"code": detail_code, "message": result.message or "post search failed"},
            )
        published_time = context.configuration.published_time.policy()
        try:
            return [post_summary(post, published_time) for post in result.data or ()]
        except PublishedTimeError as error:
            raise publication_time_http_error(error) from error

    @router.get(
        "/posts/{service}/{creator_id}/{post_id}",
        response_model=PawchivePostDetailResponse | PawchiveRevisionDetailResponse,
    )
    async def post_details(
        service: str,
        creator_id: str,
        post_id: str,
        request: Request,
        _: SessionDependency,
        revision_id: str | None = None,
    ) -> PawchivePostDetailResponse | PawchiveRevisionDetailResponse:
        result = await fetch_post(runtime(request), service, creator_id, post_id, revision_id)
        context = runtime(request)
        published_time = context.configuration.published_time.policy()
        try:
            return (
                revision_detail(result, published_time)
                if isinstance(result, Revision)
                else post_detail(result, published_time)
            )
        except PublishedTimeError as error:
            raise publication_time_http_error(error) from error

    @router.get(
        "/posts/{service}/{creator_id}/{post_id}/revisions",
        response_model=list[PawchiveRevisionSummaryResponse],
    )
    async def post_revisions(
        service: str,
        creator_id: str,
        post_id: str,
        request: Request,
        _: SessionDependency,
    ) -> list[PawchiveRevisionSummaryResponse]:
        context = runtime(request)
        try:
            with context.activate():
                async with create_pawchive_client() as client:
                    revisions = await client.list_post_revisions(service, creator_id, post_id)
                    published_time = context.configuration.published_time.policy()
                    try:
                        return [revision_summary(revision, published_time) for revision in revisions]
                    except PublishedTimeError as error:
                        raise publication_time_http_error(error) from error
        except PawchiveNotFoundError as error:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail={"code": "resource_not_found", "message": str(error)},
            ) from error
        except (PawchiveError, ValidationError) as error:
            raise upstream_error(str(error)) from error

    @router.get("/site-version", response_model=SiteVersionResponse)
    async def site_version(request: Request, _: SessionDependency) -> SiteVersionResponse:
        context = runtime(request)
        try:
            with context.activate():
                async with create_pawchive_client() as client:
                    version = await client.get_app_version()
        except PawchiveError as error:
            raise upstream_error(str(error)) from error
        return SiteVersionResponse(version=version)

    return router


async def fetch_post(
    context: RuntimeContext,
    service: str,
    creator_id: str,
    post_id: str,
    revision_id: str | None = None,
) -> Post | Revision:
    try:
        with context.activate():
            async with create_pawchive_client() as client:
                if revision_id is None:
                    return await client.get_post(service, creator_id, post_id)
                revisions = await client.list_post_revisions(service, creator_id, post_id)
        revision = next((item for item in revisions if str(item.revision_id) == revision_id), None)
        if revision is None:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail={"code": "revision_not_found", "message": "revision not found"},
            )
        return revision
    except PawchiveNotFoundError as error:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"code": "resource_not_found", "message": str(error)},
        ) from error
    except (PawchiveError, ValidationError) as error:
        raise HTTPException(
            status_code=status.HTTP_502_BAD_GATEWAY,
            detail={"code": "upstream_error", "message": str(error)},
        ) from error