# webui/openapi.yaml
openapi: 3.1.0
info:
  title: KToolBox WebUI API
  description: Authenticated REST API for the KToolBox WebUI. Selected operations also define the curated MCP tool surface
    through x-ktoolbox-mcp metadata.
  version: 1.0.0
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
paths:
...
  /api/v1/creators:
    get:
      summary: List configured creators
      operationId: list_creators
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CreatorRosterItemResponse'
                type: array
                title: Response List Creators
      description: List the project creator roster with cached Pawchive presentation names.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: list_creators
        scope: mcp:read
        safety: read
        openWorld: false
    post:
      summary: Add a creator
      operationId: add_creator
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorReference'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorReference'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Add a creator identity to the project roster.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: add_creator
        scope: mcp:write
        safety: write
        openWorld: false
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
...
  /api/v1/pawchive/creators:
    get:
      summary: Search Pawchive creators
      operationId: creators
      parameters:
      - name: creator_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - name: service
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Service
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreatorSearchItemResponse'
                title: Response Creators
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Search Pawchive for creators using an ID, name, or platform filter.
      tags:
      - pawchive
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: search_creators
        scope: mcp:read
        safety: read
        openWorld: true
  /api/v1/pawchive/posts:
    get:
      summary: Search Pawchive works
      operationId: posts
      parameters:
      - name: creator_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - name: service
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Service
      - name: query
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
      - name: offset
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PawchivePostSummaryResponse'
                title: Response Posts
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Search Pawchive for works using creator identity and text filters.
      tags:
      - pawchive
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: search_works
        scope: mcp:read
        safety: read
        openWorld: true
  /api/v1/pawchive/posts/{service}/{creator_id}/{post_id}:
    get:
      summary: Get a Pawchive work
      operationId: post_details
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: revision_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Revision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/PawchivePostDetailResponse'
                - $ref: '#/components/schemas/PawchiveRevisionDetailResponse'
                title: Response Post Details
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return one Pawchive work or a selected historical revision.
      tags:
      - pawchive
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: post_details
        scope: null
        safety: null
        openWorld: false
  /api/v1/pawchive/posts/{service}/{creator_id}/{post_id}/revisions:
    get:
      summary: List work revisions
      operationId: post_revisions
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PawchiveRevisionSummaryResponse'
                title: Response Post Revisions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return the known historical revisions for one Pawchive work.
      tags:
      - pawchive
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: list_work_revisions
        scope: mcp:read
        safety: read
        openWorld: true
...
  /api/v1/mcp/pawchive/posts/{service}/{creator_id}/{post_id}:
    get:
      tags:
      - pawchive
      summary: Get a bounded Pawchive work
      operationId: get_mcp_work
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: revision_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Revision Id
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Content
      - name: content_limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20000
          minimum: 1
          default: 12000
          title: Content Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/MCPPostResponse'
                - $ref: '#/components/schemas/MCPRevisionResponse'
                title: Response Get Mcp Work
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return one Pawchive work with content omitted by default and explicitly capped when requested.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: get_work
        scope: mcp:read
        safety: read
        openWorld: true
...
components:
  schemas:
...
    CreatorReference:
      properties:
        service:
          type: string
          minLength: 1
          title: Service
        creator_id:
          type: string
          minLength: 1
          title: Creator Id
        alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - service
      - creator_id
      title: CreatorReference
      description: A stable Pawchive creator identity stored in the project roster.
    CreatorRosterItemResponse:
      properties:
        service:
          type: string
          minLength: 1
          title: Service
        creator_id:
          type: string
          minLength: 1
          title: Creator Id
        alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
        enabled:
          type: boolean
          title: Enabled
          default: true
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        avatar:
          $ref: '#/components/schemas/MediaAssetResponse'
        banner:
          $ref: '#/components/schemas/MediaAssetResponse'
      additionalProperties: false
      type: object
      required:
      - service
      - creator_id
      - avatar
      - banner
      title: CreatorRosterItemResponse
    CreatorSearchItemResponse:
      properties:
        favorited:
          type: integer
          title: Favorited
        id:
          type: string
          minLength: 1
          title: Id
        indexed:
          anyOf:
          - type: number
          - type: string
            format: date-time
          title: Indexed
        name:
          type: string
          title: Name
        service:
          type: string
          minLength: 1
          title: Service
        updated:
          anyOf:
          - type: number
          - type: string
            format: date-time
          title: Updated
        avatar:
          $ref: '#/components/schemas/MediaAssetResponse'
        banner:
          $ref: '#/components/schemas/MediaAssetResponse'
      additionalProperties: true
      type: object
      required:
      - favorited
      - id
      - indexed
      - name
      - service
      - updated
      - avatar
      - banner
      title: CreatorSearchItemResponse
...
    FileReference:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        path:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
        deferred:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Deferred
        preview_only:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Preview Only
      additionalProperties: true
      type: object
      title: FileReference
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MediaAssetResponse:
      properties:
        kind:
          type: string
          enum:
          - avatar
          - banner
          - cover
          - attachment
          - content
          title: Kind
        thumbnail_url:
          type: string
          title: Thumbnail Url
        preview_url:
          type: string
          title: Preview Url
        original_url:
          type: string
          title: Original Url
      type: object
      required:
      - kind
      - thumbnail_url
      - preview_url
      - original_url
      title: MediaAssetResponse
...
    PawchivePostDetailResponse:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        user:
          type: string
          minLength: 1
          title: User
        service:
          type: string
          minLength: 1
          title: Service
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        substring:
          anyOf:
          - type: string
          - type: 'null'
          title: Substring
        embed:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Embed
        shared_file:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Shared File
        added:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Added
        published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published
        edited:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Edited
        file:
          anyOf:
          - $ref: '#/components/schemas/FileReference'
          - type: 'null'
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/FileReference'
            type: array
          - type: 'null'
          title: Attachments
        poll:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Poll
        captions:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Captions
        tags:
          anyOf:
          - type: string
          - items: {}
            type: array
          - type: 'null'
          title: Tags
        origin:
          anyOf:
          - type: string
          - type: 'null'
          title: Origin
        preview_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview State
        has_full:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Full
        preview_attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Preview Attempts
        detail_fetched:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Detail Fetched
        import_size_cap_gb:
          anyOf:
          - type: number
          - type: 'null'
          title: Import Size Cap Gb
        next:
          anyOf:
          - type: string
          - type: 'null'
          title: Next
        prev:
          anyOf:
          - type: string
          - type: 'null'
          title: Prev
        cover:
          anyOf:
          - $ref: '#/components/schemas/MediaAssetResponse'
          - type: 'null'
        effective_published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Effective Published
        published_service_timezone:
          type: string
          title: Published Service Timezone
        published_target_timezone:
          type: string
          title: Published Target Timezone
        media:
          items:
            $ref: '#/components/schemas/MediaAssetResponse'
          type: array
          title: Media
      additionalProperties: true
      type: object
      required:
      - id
      - user
      - service
      - published_service_timezone
      - published_target_timezone
      title: PawchivePostDetailResponse
    PawchivePostSummaryResponse:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        user:
          type: string
          minLength: 1
          title: User
        service:
          type: string
          minLength: 1
          title: Service
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        substring:
          anyOf:
          - type: string
          - type: 'null'
          title: Substring
        embed:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Embed
        shared_file:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Shared File
        added:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Added
        published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published
        edited:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Edited
        file:
          anyOf:
          - $ref: '#/components/schemas/FileReference'
          - type: 'null'
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/FileReference'
            type: array
          - type: 'null'
          title: Attachments
        poll:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Poll
        captions:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Captions
        tags:
          anyOf:
          - type: string
          - items: {}
            type: array
          - type: 'null'
          title: Tags
        origin:
          anyOf:
          - type: string
          - type: 'null'
          title: Origin
        preview_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview State
        has_full:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Full
        preview_attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Preview Attempts
        detail_fetched:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Detail Fetched
        import_size_cap_gb:
          anyOf:
          - type: number
          - type: 'null'
          title: Import Size Cap Gb
        next:
          anyOf:
          - type: string
          - type: 'null'
          title: Next
        prev:
          anyOf:
          - type: string
          - type: 'null'
          title: Prev
        cover:
          anyOf:
          - $ref: '#/components/schemas/MediaAssetResponse'
          - type: 'null'
        effective_published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Effective Published
        published_service_timezone:
          type: string
          title: Published Service Timezone
        published_target_timezone:
          type: string
          title: Published Target Timezone
      additionalProperties: true
      type: object
      required:
      - id
      - user
      - service
      - published_service_timezone
      - published_target_timezone
      title: PawchivePostSummaryResponse
    PawchiveRevisionDetailResponse:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        user:
          type: string
          minLength: 1
          title: User
        service:
          type: string
          minLength: 1
          title: Service
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        substring:
          anyOf:
          - type: string
          - type: 'null'
          title: Substring
        embed:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Embed
        shared_file:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Shared File
        added:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Added
        published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published
        edited:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Edited
        file:
          anyOf:
          - $ref: '#/components/schemas/FileReference'
          - type: 'null'
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/FileReference'
            type: array
          - type: 'null'
          title: Attachments
        poll:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Poll
        captions:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Captions
        tags:
          anyOf:
          - type: string
          - items: {}
            type: array
          - type: 'null'
          title: Tags
        origin:
          anyOf:
          - type: string
          - type: 'null'
          title: Origin
        preview_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview State
        has_full:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Full
        preview_attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Preview Attempts
        detail_fetched:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Detail Fetched
        import_size_cap_gb:
          anyOf:
          - type: number
          - type: 'null'
          title: Import Size Cap Gb
        next:
          anyOf:
          - type: string
          - type: 'null'
          title: Next
        prev:
          anyOf:
          - type: string
          - type: 'null'
          title: Prev
        revision_id:
          type: integer
          title: Revision Id
        cover:
          anyOf:
          - $ref: '#/components/schemas/MediaAssetResponse'
          - type: 'null'
        effective_published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Effective Published
        published_service_timezone:
          type: string
          title: Published Service Timezone
        published_target_timezone:
          type: string
          title: Published Target Timezone
        media:
          items:
            $ref: '#/components/schemas/MediaAssetResponse'
          type: array
          title: Media
      additionalProperties: true
      type: object
      required:
      - id
      - user
      - service
      - revision_id
      - published_service_timezone
      - published_target_timezone
      title: PawchiveRevisionDetailResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    WebUISession:
      type: apiKey
      in: cookie
      name: ktoolbox_session
      description: Opaque browser session created by the login operation.
...